requirements.txt
crewai 
langchain-google-genai 
pyyaml 
arxiv 
requests